Extract shared interfaces to `lib/interfaces.ts`.
